fix(collections): include crate in Docker image and cast collected amount
Two bugs caught during staging validation:

- core/Dockerfile: add COPY collections/ — the new core/collections crate was
  missing from the image's per-crate copy list, so the core image failed to build
  (failed to load manifest for workspace member /app/collections).

- core/collections repository: collected_amount decoded SUM(amount_minor) as i64,
  but Postgres SUM(bigint) returns NUMERIC, causing a 500 on GET collection by id
  against a real database. Cast COALESCE(SUM(...),0)::bigint so it decodes as i64.
  The in-memory unit tests bypassed sqlx and missed this; staging caught it.

cargo build + collections invariants (11/11) green.

+++ core/collections/src/repository.rs
@@ -526,8 +526,10 @@ impl CollectionRepository for PostgresCollectionRepository {
         &self,
         collection_id: CollectionId,
     ) -> Result<i64, CollectionError> {
+        // SUM(bigint) returns NUMERIC in Postgres — cast back to bigint so it
+        // decodes as i64.
         let sum: Option<i64> = sqlx::query_scalar(
-            "SELECT COALESCE(SUM(amount_minor), 0) FROM collection_shares \
+            "SELECT COALESCE(SUM(amount_minor), 0)::bigint FROM collection_shares \
              WHERE collection_id = $1 AND status = 'PAID'",
         )
         .bind(collection_id.as_uuid())
